+++ src/constants/MoveHistory/EventTemplates/UpdateMTOShipment/createMTOShipmentAgent.jsx
@@ -2,10 +2,11 @@ import { formatMoveHistoryAgent } from 'utils/formatters';
 import d from 'constants/MoveHistory/UIDisplay/DetailsTypes';
 import t from 'constants/MoveHistory/Database/Tables';
 import a from 'constants/MoveHistory/Database/Actions';
+import o from 'constants/MoveHistory/UIDisplay/Operations';
 
 export default {
   action: a.INSERT,
-  eventName: '*',
+  eventName: o.updateMTOShipment,
   tableName: t.mto_agents,
   detailsType: d.LABELED,
   getEventNameDisplay: () => 'Updated shipment',
